# Enable SIO_LOOPBACK_FAST_PATH on Windows