@jtattermusch 2018-04-03
Are we seeing any performance improvements thanks to this change?