@grpc-testing 2018-04-03
```
[trickle] No significant performance differences
```