@ara-ayvazyan 2018-04-03
Yes, significant. For one of our basic tests it shows over 2x throughput improvement (Windows 10).