@ara-ayvazyan 2018-04-04
This seem to reveal a high CPU perf issue in [`gpr_log`](https://github.com/grpc/grpc/blob/master/src/core/lib/gpr/log_windows.cc#L54) when using `vsnprintf_s`. Will send a PR for that one.