@ara-ayvazyan 2018-04-05
Reusing read buffers and adjusting the size also helps (#14955 ).