@ncteisen 2018-04-19
Can you share performance data? Just filed #15125 for standardized Windows perf tooling