@ara-ayvazyan 2018-04-19
I don't have detailed data at the moment, but the ratio is around 2x.