@ncteisen 2018-04-23
2X what? QPS? Throughput? CPU time or Wall time (I assume not 😀)

How big were the payloads? etc etc etc.

Getting before/after data from microbenchmarks or our synthetic system would be invaluable because we are all used to looking at those numbers.