@ara-ayvazyan 2018-04-26
Improvement is in latency and CPU. Below numbers are taken from a test that was running for 1 min sending and receiving a blob of specified size containing random data and measuring the latency.

Without `SIO_LOOPBACK_FAST_PATH`:

| QPS | Payload | 50% | 90% | 95% | 99% |
|:-----:|:-----:|:-----:|-----|-----|-----|
| 1000 | 1KB | 1229μs | 1638μs | 1744μs | 1957μs |
| 1000 | 100KB | 17814μs | 26750μs | 29861μs | 40768μs |
| 10000 | 10KB | 15504μs | 19842μs | 21428μs | 27766μs |

With `SIO_LOOPBACK_FAST_PATH`:

| QPS | Payload | 50% | 90% | 95% | 99% |
|:-----:|:-----:|:-----:|-----|-----|-----|
| 1000 | 1KB | 897μs | 1340μs | 1455μs | 1663μs |
| 1000 | 100KB | 6715μs | 8827μs | 9683μs | 11070μs |
| 10000 | 10KB | 9442μs | 13040μs | 13757μs | 15283μs |