@ncteisen 2018-04-26
Awesome, thank you for the performance data!

@nicolasnoble or @sreecha, can you take a look at this. It looks good to me, but I have no Windows knowledge.