@ncteisen 2018-04-26
@ara-ayvazyan, this looks good. Can you rebase?