@grpc-testing 2018-04-27
```
[trickle] No significant performance differences
```